+++ src/main/java/com/tobiasdiez/easybind/select/NestedSelectionElement.java
@@ -1,0 +1,8 @@
+package com.tobiasdiez.easybind.select;
+
+interface NestedSelectionElement<T, U> {
+    void connect(T baseVal);
+    void disconnect();
+    boolean isConnected();
+    U getValue();
+}