+++ src/main/java/com/tobiasdiez/easybind/select/NestedSelectionElementFactory.java
@@ -1,0 +1,6 @@
+package com.tobiasdiez.easybind.select;
+
+@FunctionalInterface
+interface NestedSelectionElementFactory<T, U> {
+    NestedSelectionElement<T, U> create(Runnable invalidationCallback);
+}