+++ src/main/java/com/tobiasdiez/easybind/select/RootSelectedBuilder.java
@@ -1,0 +1,19 @@
+package com.tobiasdiez.easybind.select;
+
+import javafx.beans.value.ObservableValue;
+
+import com.tobiasdiez.easybind.monadic.MonadicBinding;
+
+class RootSelectedBuilder<T> implements ParentSelectedBuilder<T> {
+    private final ObservableValue<T> root;
+
+    public RootSelectedBuilder(ObservableValue<T> root) {
+        this.root = root;
+    }
+
+    @Override
+    public <U> MonadicBinding<U> create(
+            NestedSelectionElementFactory<T, U> nestedSelectionFactory) {
+        return new SelectObjectBinding<T, U>(root, nestedSelectionFactory);
+    }
+}